Wrapping up the loose ends. Jenny pointed out that the wizard should allow one to skip the Advanced step altogether, as it is optional. I found that while the wizard code is easily instructed to display the finish button before the last step, it covered the Next and finish buttons with a DIV that intercepted the clicks on either one. That had to be moved out of the way so those buttons would receive the click event and fire the appropriate handler. I also had to make the finish button handler check the current page validation, just like a click on the next button does.
Other changes in this commit are small name changes to a couple of block device mapping fields, to match what models/launchconfig.js is expecting. Now firing the sync event in that model and sending the AJAX request. This test box isn't communicating with the live facility, so I'll have to arrange that to make sure it will do its job.