Making the Save button behave like the other dialogs, in that it is disabled until validation passes. Tricky, because the tag editor UI widget handles the validation of the tags - not the dialog view.
Made the tag editor send up a 'validation_change' event when a tag is validated, pass or faili, and provide the model in question. Each tag model already contains the attribute 'tag_is_invalid' for the benefit of the view iterator displaying the error message right beneath each invalid tag, instead of only at the bottom of the list (EUCA-6173), so get that value and track the invalid ones in a collection until they're all fixed before enabling the Save button.